lnrc-dashboard/bot/requirements.txt
python-telegram-bot==13.15
python-dotenv
web3
